removed strong type comparison between ids on image deletion
newer PHP/SQL version will return integer instead of string for integer fields

+++ modules/Vtiger/models/Record.php
@@ -592,7 +592,7 @@ class Vtiger_Record_Model extends Vtiger_Base_Model {
 		$checkResult = $db->pquery('SELECT crmid FROM vtiger_seattachmentsrel WHERE attachmentsid = ?', array($imageId));
 		$crmId = $db->query_result($checkResult, 0, 'crmid');
 
-		if ($this->getId() === $crmId) {
+		if (!empty($imageId) && $this->getId() == $crmId) {
 			$db->pquery('DELETE FROM vtiger_attachments WHERE attachmentsid = ?', array($imageId));
 			$db->pquery('DELETE FROM vtiger_seattachmentsrel WHERE attachmentsid = ?', array($imageId));
 			return true;
